requirements.txt
psycopg2-binary
django
django-crispy-forms
requests